fix for #114
I'm not sure this is the best way to do it, but I think it makes sense. Essentially, it ignores the block when the type is already error on a block call.

also, the test should be part of the core test suite, but I couldn't figure out how to make a failing test for this case in that suite.