fixes #52 by raising an inference error when types arent compat
note: I ran into an issue where deferred inference assumes that you are inferring as an expression,
which isn't always true for a rescue statement

so I put in a ivar to cache the initial expression value--I feel like this should be done in a more general way, but my main goal was to fix the bug.

I also noticed that there is a lot of variation in how the infer(typer, exp) methods determined whether they are cached or not. It' be nice to clean that up. I might do some of that next weekend.