add some tests around assignment types, 1 working 2 pending
0ed0ed8a08d98f2ed8fcf69934bf76e60c5f7d2c made it so that

  a = b = c

is equivalent to

  b = c
  a = c

in terms of type inference.

This commit adds some tests around assignment/return edge cases
with primitives, since they must be explicitly converted in
bytecode.

A couple of them are pending. Essentially the problem is that if
the last expression in a method is of type X, but the method is
supposed to return Y where X can be widened to Y, the widening bit
isn't generated