fix inference of begin;rescue with raise in rescue
the parser fills in the node list for the else portion of a rescue, even when it doesn't exist. The code that determines the type of a begin rescue assumes it will be nil if it doesn't exist.

we need to get the parser to not include it, but the parser needs to be buildable first, and this bug prevented mmeta from compiling