fix gem task definition ordering
Rubygems PackageTask defines task :gem as pointing to the file it generates. I was adding bootstrap requirements after, which meant that building jars would happen _after_ the gem was completed. That sucks. This reordering fixes it by predefining the gem dependencies s.t. when the PackageTask is defined, gem already has dependencies and they are run first

https://github.com/rubygems/rubygems/blob/master/lib/rubygems/package_task.rb#L109