fix BlocksTest#test_block_with_method_def
we were inferring block args too early and more importantly, introducing a scope around them. That put those args in a scope that was captured prematurely

Because they were captured, the compiler would think they were fields in a binding, even if the block was inlined as part of a macro, causing bytecode for an invalid field to be generated.