add option to compile macros to separate destination
Macros can now be compiled to a separate folder. What happens is that now, you can specify --macro-dest <>, and your macro code, instead of living with other generated artifacts, is instead compiled there. By default, it's added to the macro classpath, when specified, so you can do separate compilations.

In addition, the compile only annotations are now only on files dropped in the macro destination. Instead of macros causing the Extensions annotation to live on the class they extend, the annotation goes on a specially named class that exists only for the purpose of looking up extensions. It has the name <class-being-extended>. The macro loader code, now looks up this on the macro classpath. By default, that's the same as the regular classpath, but if you set the macro-dest, you can use that instead.

This should make the class files that mirahc generates more friendly to other tools. You shouldn't have to have mirahc on your path to be able to build against mirah created class files.