skip unresolved imports when generating macro templates
We could instead wait until the imports are resolved before generating macros. I put a little effort into that, but ran into issues. This at least fixes the current pain.

I'm thinking when extension methods become user land features we'll have to do some work with how macros and imports interact which might be a good time to revisit this if we don't address it earlier.