Fix metacircularity of the parser.
There were two major bits.
One, resolved types are compared by ID so that meant that two identical types that some how ended up with the same name would not be equal. :(
     resolved that by comparing the string versions. Not great, but got us past it.
Two, casts are now more restrictive, so in order to cast down, the cast has to be superclass -> subclass, not any class to any other class. So needed a cast to object first in a few places

Also, revamped some tests and how all of the build bits fit together.